refactor: order_table 을 eat_in_order 패키지 하위로 이동 - order_table 은 eat_in_order 와 항상 함께 사용됨
