+++ src/main/java/kitchenpos/order/eat_in_order/application/OrderTableService.java
@@ -1,9 +1,9 @@
-package kitchenpos.order_table.application;
+package kitchenpos.order.eat_in_order.application;
 
-import kitchenpos.order.eat_in_order.domain.EatInOrderRepository;
-import kitchenpos.order.eat_in_order.domain.OrderStatus;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.domain.OrderTableRepository;
+import kitchenpos.order.eat_in_order.domain.eat_in_order.EatInOrderRepository;
+import kitchenpos.order.eat_in_order.domain.eat_in_order.OrderStatus;
+import kitchenpos.order.eat_in_order.domain.order_table.OrderTable;
+import kitchenpos.order.eat_in_order.domain.order_table.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
