+++ src/main/java/kitchenpos/order/eat_in_order/domain/eat_in_order/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.eat_in_order.domain;
+package kitchenpos.order.eat_in_order.domain.eat_in_order;
 
 public enum OrderStatus {
     WAITING, ACCEPTED, SERVED, COMPLETED
