+++ src/main/java/kitchenpos/order/eat_in_order/infra/JpaEatInOrderRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.eat_in_order.infra;
 
-import kitchenpos.order.eat_in_order.domain.EatInOrder;
-import kitchenpos.order.eat_in_order.domain.EatInOrderRepository;
+import kitchenpos.order.eat_in_order.domain.eat_in_order.EatInOrder;
+import kitchenpos.order.eat_in_order.domain.eat_in_order.EatInOrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
