+++ src/main/java/kitchenpos/order/eat_in_order/infra/JpaOrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.order_table.infra;
+package kitchenpos.order.eat_in_order.infra;
 
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.domain.OrderTableRepository;
+import kitchenpos.order.eat_in_order.domain.order_table.OrderTable;
+import kitchenpos.order.eat_in_order.domain.order_table.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
